refactor: export public API explicitly
Closes #99
