+++ packages/cbl/lib/cbl.dart
@@ -1,28 +1,8 @@
-export 'src/couchbase_lite.dart'
-    hide debugCouchbaseLiteIsInitialized, debugRefCounted;
-export 'src/database.dart' hide DatabaseImpl;
-export 'src/document/array.dart' hide ArrayImpl, MutableArrayImpl;
-export 'src/document/blob.dart' hide BlobImpl, BlobImplSetter;
-export 'src/document/dictionary.dart'
-    hide DictionaryImpl, MutableDictionaryImpl;
-export 'src/document/document.dart'
-    hide
-        DocumentImpl,
-        MutableDocumentImpl,
-        DocumentMContext,
-        DocumentEncoderContext;
-export 'src/document/fragment.dart'
-    hide FragmentImpl, MutableFragmentImpl, DocumentFragmentImpl;
+export 'src/couchbase_lite.dart';
+export 'src/database.dart';
+export 'src/document.dart';
 export 'src/errors.dart' hide translateCBLErrorException, CBLErrorExceptionExt;
 export 'src/log.dart';
-export 'src/query.dart' hide QueryImpl;
-export 'src/replication/authenticator.dart';
-export 'src/replication/configuration.dart';
-export 'src/replication/conflict.dart' hide ConflictImpl;
-export 'src/replication/conflict_resolver.dart';
-export 'src/replication/document_replication.dart'
-    hide DocumentReplicationImpl, ReplicatedDocumentImpl;
-export 'src/replication/endpoint.dart';
-export 'src/replication/replicator.dart' hide ReplicatorImpl;
-export 'src/replication/replicator_change.dart' hide ReplicatorChangeImpl;
+export 'src/query.dart';
+export 'src/replication.dart';
 export 'src/support/resource.dart' show Resource, ClosableResource;
