+++ packages/cbl/lib/src/document.dart
@@ -1,0 +1,21 @@
+export 'document/array.dart'
+    show Array, ArrayInterface, MutableArray, MutableArrayInterface;
+export 'document/blob.dart' show Blob;
+export 'document/dictionary.dart'
+    show
+        Dictionary,
+        DictionaryInterface,
+        MutableDictionary,
+        MutableDictionaryInterface;
+export 'document/document.dart' show Document, MutableDocument;
+export 'document/fragment.dart'
+    show
+        ArrayFragment,
+        DictionaryFragment,
+        DocumentFragment,
+        Fragment,
+        FragmentInterface,
+        MutableArrayFragment,
+        MutableFragment,
+        MutableFragmentInterface,
+        MutableDictionaryFragment;
