+++ packages/cbl/lib/src/document/blob.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 import 'package:cbl_ffi/cbl_ffi.dart';
 import 'package:collection/collection.dart';
 
-import '../database.dart';
+import '../database/database.dart';
 import '../fleece/encoder.dart';
 import '../fleece/fleece.dart';
 import '../fleece/fleece.dart' as fl;
