+++ packages/cbl/lib/src/replication.dart
@@ -1,0 +1,21 @@
+export 'replication/authenticator.dart'
+    show Authenticator, BasicAuthenticator, SessionAuthenticator;
+export 'replication/configuration.dart'
+    show
+        ReplicatorConfiguration,
+        DocumentFlag,
+        ReplicatorType,
+        ReplicationFilter;
+export 'replication/conflict.dart' show Conflict;
+export 'replication/conflict_resolver.dart'
+    show ConflictResolver, ConflictResolverFunction;
+export 'replication/document_replication.dart'
+    show DocumentReplication, ReplicatedDocument;
+export 'replication/endpoint.dart' show Endpoint, UrlEndpoint;
+export 'replication/replicator.dart'
+    show
+        Replicator,
+        ReplicatorProgress,
+        ReplicatorActivityLevel,
+        ReplicatorStatus;
+export 'replication/replicator_change.dart' show ReplicatorChange;
